refactor(settings): CloseConfirmDialog + 업데이트 헬퍼 분리
대형 파일 정리 9번째(Phase I). settings.py(1180줄)에서 SettingsDialog와
독립적인 CloseConfirmDialog(앱 종료 확인 다이얼로그, main_window.py가
직접 재사용)와 업데이트 관련 소품 3개(UpdatePackageActionText,
update_package_action_text, UpdateCheckerThread)를 분리.

- settings_close_confirm_dialog.py (신규): CloseConfirmDialog
- settings_update_helpers.py (신규): UpdatePackageActionText(dataclass),
  update_package_action_text, UpdateCheckerThread
- settings.py: 1180 -> 1078줄 (재수입)

main_window.py의 `from src.ui.dialogs.settings import CloseConfirmDialog,
SettingsDialog`는 변경 없이 그대로 동작(파사드 재수입).

검증: 관련 pytest 10 passed. 전체 pytest 1809 passed / 1 known-flaky
제외. py_compile 통과. CloseConfirmDialog/SettingsDialog 오프스크린
오픈 확인. 앱 스모크 성공.
